deepseek/mqtt/requirements.txt
aiohttp>=3.8.0
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
sqlalchemy>=1.4.0
alembic>=1.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pyyaml>=6.0
websockets>=10.0
asyncio-mqtt>=0.13.0
psutil>=5.9.0
python-multipart>=0.0.5
cryptography>=3.4
python-dotenv>=0.19.0
pydantic>=1.10.0